requirements.txt
# Python 3.10.4

opencv-python==4.6.0
numpy==1.23.2
pillow==9.2.0
imutils==0.5.4
scipy==1.9.0